requirements.txt
SoundFile==0.10.3.post1
tensorflow_datasets==4.4.0
nltk==3.6.4
numpy~=1.19.2
sentencepiece==0.1.96
tqdm==4.62.1
librosa==0.8.1
PyYAML==5.4.1
Pillow==8.3.2
black==21.7b0
flake8==3.9.2
sounddevice==0.4.3
pybind11
fastwer
cached_properties

# extra=tf2.5-gpu
tensorflow-gpu~=2.5.1
tensorflow-text~=2.5.0
tensorflow-io~=0.18.0